fix: address code review issues in email templates feature
- Use validateAndParseId() for [id] route params instead of manual parseInt
- Fix hook mutations to re-throw non-auth errors so page catch blocks fire
- Add public /api/staff/email-templates endpoint for active templates (staff+admin accessible)
- Split useEmailTemplates into fetchActiveTemplates/fetchAllTemplates matching volunteer-types pattern
- Remove broken JSX comment fragment in compose-modal
- Change deleteTemplate from soft-delete to hard-delete; UI copy already said permanently
- Extract page client logic to email-templates-settings-client.tsx following settings page pattern
